create recording time tag part, should inherit TimelinePart to get the right position.
